# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Barrio Fino Peluqueria API",
        "description": "API documentation for Barrio Fino Peluqueria application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/usuarios": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "description": "Retrieve a list of all users",
                "operationId": "306f2624b9d148337d4bed240d2dceb1",
                "responses": {
                    "200": {
                        "description": "Successfull operation"
                    },
                    "404": {
                        "description": "No users found"
                    }
                }
            }
        },
        "/api/v1/usuarios/store": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "description": "Create a new user with the provided information",
                "operationId": "55b157d1e3d2585ac4a6a1a2f6f8cfeb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role",
                                    "phone_number",
                                    "photo_url"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "doe@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "cliente"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "photo_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://example.com/photo.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                }
            }
        },
        "/api/v1/usuarios/search/{nombre}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Search users by nombre",
                "operationId": "b892e0d28bc71b555c4d4629db4dcee0",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "path",
                        "description": "Nombre of the users to search for",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "John"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users found successfully"
                    },
                    "404": {
                        "description": "No users found"
                    }
                }
            }
        },
        "/api/v1/usuarios/barberos": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all barbers",
                "operationId": "a190844cfa133ad2209f9ebccc6e67c8",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "David"
                                                    },
                                                    "role": {
                                                        "type": "string",
                                                        "example": "barbero"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No barbers found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No barbers found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}